Fucking submodule bull shit

+++ src/dev/namtx/leetcode/daily/MinimumAverageDifference.java
@@ -7,16 +7,16 @@ public class MinimumAverageDifference {
     }
 
     public int minimumAverageDifference(int[] nums) {
-        int[] prefixSum = new int[nums.length];
+        long[] prefixSum = new long[nums.length];
         prefixSum[0] = nums[0];
         for (int i = 1; i < nums.length; i++) {
             prefixSum[i] = nums[i] + prefixSum[i - 1];
         }
         int ans = 0;
-        int min = Integer.MAX_VALUE;
+        long min = Long.MAX_VALUE;
         for (int i = 0; i < nums.length - 1; i++) {
-            int firstAverage = prefixSum[i] / (i + 1);
-            int lastAverage = (prefixSum[nums.length - 1] - prefixSum[i]) / (nums.length - i - 1);
+            long firstAverage = prefixSum[i] / (i + 1);
+            long lastAverage = (prefixSum[nums.length - 1] - prefixSum[i]) / (nums.length - i - 1);
             System.out.println(Math.abs(firstAverage - lastAverage));
             if (min > Math.abs(firstAverage - lastAverage)) {
                 min = Math.abs(firstAverage - lastAverage);
